feat: add casts to channels token details
