+++ views/token-detail/dummy.ts
@@ -281,181 +281,6 @@ const generateData = (period: Period, isSecondary?: boolean) => {
   return data;
 };
 
-const comments: ILBComment[] = [
-  {
-    comment:
-      'Rolling out channel specific user rankings think of these as customizable channel power badges - an essential lego for decentralized channels - rankings for 100 channels live, 500+ by eow - ongoing eyeballing/feedback from channel owners and @automod @jrgi - soon @neynar makes it easy to consume for devs/frames more ⭐️',
-    createdAt: '2024-06-08T00:00:00Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fhjfh7wdws9',
-      name: 'Sahil',
-      username: 'sahil',
-    },
-    id: 'fhjfh7hkjsu',
-    images: ['https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718804332/image_54_u75dgm.jpg'],
-    likes: 5700,
-    recasts: 12,
-    replies: [
-      {
-        comment:
-          'Excited to announce the integration of new features in our platform! 🎉 Continuous improvement and feedback from our community has been invaluable. Check out the new update and let us know your thoughts! #productupdate #community',
-        createdAt: '2024-06-10T12:34:56Z',
-        createdBy: {
-          avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-          id: 'fjkfj83hdsn',
-          name: 'Ananya',
-          username: 'ananya',
-        },
-        id: 'fjkfj83hdsn1',
-        images: [],
-        likes: 3200,
-        recasts: 20,
-        replies: [],
-        updatedAt: '2024-06-10T12:34:56Z',
-      },
-      {
-        comment:
-          'Excited to announce the integration of new features in our platform! 🎉 Continuous improvement and feedback from our community has been invaluable. Check out the new update and let us know your thoughts! #productupdate #community',
-        createdAt: '2024-06-10T12:34:56Z',
-        createdBy: {
-          avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-          id: 'fjkfj83hdsn',
-          name: 'Ananya',
-          username: 'ananya',
-        },
-        id: 'fjkfj83hdsn1',
-        images: [],
-        likes: 3200,
-        recasts: 20,
-        replies: [],
-        updatedAt: '2024-06-10T12:34:56Z',
-      },
-      {
-        comment:
-          'Excited to announce the integration of new features in our platform! 🎉 Continuous improvement and feedback from our community has been invaluable. Check out the new update and let us know your thoughts! #productupdate #community',
-
-        createdAt: '2024-06-10T12:34:56Z',
-        createdBy: {
-          avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-          id: 'fjkfj83hdsn',
-          name: 'Ananya',
-          username: 'ananya',
-        },
-        id: 'fjkfj83hdsn1',
-        images: [],
-        likes: 3200,
-        recasts: 20,
-        replies: [],
-        updatedAt: '2024-06-10T12:34:56Z',
-      },
-      {
-        comment:
-          'Excited to announce the integration of new features in our platform! 🎉 Continuous improvement and feedback from our community has been invaluable. Check out the new update and let us know your thoughts! #productupdate #community',
-
-        createdAt: '2024-06-10T12:34:56Z',
-        createdBy: {
-          avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-          id: 'fjkfj83hdsn',
-          name: 'Ananya',
-          username: 'ananya',
-        },
-        id: 'fjkfj83hdsn1',
-        images: [],
-        likes: 3200,
-        recasts: 20,
-        replies: [],
-        updatedAt: '2024-06-10T12:34:56Z',
-      },
-    ],
-    updatedAt: '2024-06-08T00:00:00Z',
-  },
-  {
-    comment:
-      'Excited to announce the integration of new features in our platform! 🎉 Continuous improvement and feedback from our community has been invaluable. Check out the new update and let us know your thoughts! #productupdate #community',
-
-    createdAt: '2024-06-10T12:34:56Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fjkfj83hdsn',
-      name: 'Ananya',
-      username: 'ananya',
-    },
-    id: 'fjkfj83hdsn1',
-    images: [],
-    likes: 3200,
-    recasts: 20,
-    replies: [],
-    updatedAt: '2024-06-10T12:34:56Z',
-  },
-  {
-    comment:
-      'Join us for our upcoming webinar on decentralized applications! 🚀 Learn from industry experts and get insights into the future of decentralized tech. Register now! #webinar #decentralization Join us for our upcoming webinar on decentralized applications! 🚀 Learn from industry experts and get insights into the future of decentralized tech. Register now! #webinar #decentralization',
-    createdAt: '2024-06-12T08:20:15Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fjd8392kdfs',
-      name: 'Ravi',
-      username: 'ravi',
-    },
-    id: 'fjd8392kdfs2',
-    images: [],
-    likes: 4400,
-    recasts: 15,
-    replies: [],
-    updatedAt: '2024-06-12T08:20:15Z',
-  },
-  {
-    comment: 'Our team is working tirelessly to bring you the best experience. Big shoutout to everyone involved! 🙌 Stay tuned for more updates. #teamwork #development',
-    createdAt: '2024-06-14T14:45:30Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fj8390wjsdf',
-      name: 'Maya',
-      username: 'maya',
-    },
-    id: 'fj8390wjsdf3',
-    images: [],
-    likes: 5100,
-    recasts: 18,
-    replies: [],
-    updatedAt: '2024-06-14T14:45:30Z',
-  },
-  {
-    comment: 'Just released a new tutorial on how to get started with decentralized finance (DeFi). Check it out and start your DeFi journey today! 💡 #DeFi #tutorial',
-    createdAt: '2024-06-16T10:15:45Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fji3490kfns',
-      name: 'Aarav',
-      username: 'aarav',
-    },
-    id: 'fji3490kfns4',
-    images: [],
-    likes: 6000,
-    recasts: 22,
-    replies: [],
-    updatedAt: '2024-06-16T10:15:45Z',
-  },
-  {
-    comment:
-      'Proud to announce that we have reached 10,000 active users! 🎉 Thank you to our amazing community for the support. Here’s to many more milestones! #milestone #community Proud to announce that we have reached 10,000 active users! 🎉 Thank you to our amazing community for the support. Here’s to many more milestones! #milestone #communit Proud to announce that we have reached 10,000 active users! 🎉 Thank you to our amazing community for the support. Here’s to many more milestones! #milestone #community',
-    createdAt: '2024-06-18T09:00:00Z',
-    createdBy: {
-      avatar: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg',
-      id: 'fj9384jkdks',
-      name: 'Zara',
-      username: 'zara',
-    },
-    id: 'fj9384jkdks5',
-    images: [],
-    likes: 7500,
-    recasts: 30,
-    replies: [],
-    updatedAt: '2024-06-18T09:00:00Z',
-  },
-];
-
 const holders = [
   { name: 'drippy.eth', amount: 95100000, avatarURL: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723895/Avatar_khczwg.jpg' },
   { name: 'Evans', amount: 25700000, avatarURL: 'https://res.cloudinary.com/dxnd4k222/image/upload/fl_preserve_transparency/v1718723894/Avatar_1_rg5wgw.jpg' },
@@ -511,4 +336,4 @@ const leaderboardData: TableItem[] = [
   },
 ];
 
-export { tokenDetailData, transactionsData, holdingsData, periods, generateData, comments, holders, leaderboardData };
+export { tokenDetailData, transactionsData, holdingsData, periods, generateData, holders, leaderboardData };
